+++ src/test/scala/TableTest.scala
@@ -301,51 +301,4 @@ class TableTest extends FunSuite {
       thrown.getMessage === "@type of table is not 'Table' - country_slice.csv.@type"
     )
   }
-
-  test(
-    "sets scala test for Array[Any] and List[Any]"
-  ) {
-    var exampleSet: Set[List[Any]] = Set()
-    exampleSet += List[Any](12, 1, 4, "Abcd")
-    exampleSet += List[Any](12, 1, 4, "Abcd")
-    exampleSet += List[Any](12, 1, 4, "Abcd")
-
-    assert(exampleSet.size == 1)
-
-    var exampleSet2: Set[Array[Any]] = Set()
-
-    exampleSet2 += Array[Any](12, 1, 4, "Abcd")
-    exampleSet2 += Array[Any](12, 1, 4, "Abcd")
-    exampleSet2 += Array[Any](12, 1, 4, "Abcd")
-    assert(
-      exampleSet2.size == 3
-    )
-
-    var exampleSet3: Set[List[Any]] = Set()
-    val zone = ZoneId.of("UTC+1")
-    ZonedDateTime.of(1947, 8, 15, 12, 12, 12, 12, zone)
-    exampleSet3 += List[Any](
-      12,
-      1,
-      4,
-      "Abcd",
-      ZonedDateTime.of(1947, 8, 15, 12, 12, 12, 12, zone)
-    )
-    exampleSet3 += List[Any](
-      12,
-      1,
-      4,
-      "Abcd",
-      ZonedDateTime.of(1947, 8, 15, 12, 12, 12, 12, zone)
-    )
-    exampleSet3 += List[Any](
-      12,
-      1,
-      4,
-      "Abcd",
-      ZonedDateTime.of(1947, 8, 15, 12, 12, 12, 12, zone)
-    )
-
-    assert(exampleSet3.size == 1)
-  }
 }
